% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_supported_language.R
\name{check_for_supported_language}
\alias{check_for_supported_language}
\title{Internal function to validate supported languages}
\usage{
check_for_supported_language(language)
}
\arguments{
\item{language}{A character string specifying the language code to be validated.}
}
\value{
This function does not return a value. It either silently passes or throws an error if the language is unsupported.
}
\description{
This function checks if the provided language is supported by the World Bank API.
If the language is not supported, it throws an error listing the available languages.
}
\details{
This is an internal helper function that checks the language against the list of supported languages
from the World Bank API. If the language is not supported, it throws an error using the \code{cli_abort()} function.
}
\seealso{
\code{\link[=list_supported_languages]{list_supported_languages()}} for obtaining the list of supported languages.
}
\keyword{internal}
