% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_indicator.R
\name{download_indicator}
\alias{download_indicator}
\title{Download World Bank indicator data for specific countries}
\usage{
download_indicator(
  countries,
  indicator,
  start_date = NULL,
  end_date = NULL,
  language = "en",
  progress = TRUE
)
}
\arguments{
\item{countries}{A character vector of ISO 2-country codes, or \code{"all"} to retrieve data for all countries.}

\item{indicator}{A character string specifying the World Bank indicator to download (e.g., "NY.GDP.PCAP.KD").}

\item{start_date}{Optional. The starting date for the data, either as a year (e.g., \code{2010}) or a specific month (e.g., \code{"2012M01"}).}

\item{end_date}{Optional. The ending date for the data, either as a year (e.g., \code{2020}) or a specific month (e.g., \code{"2012M05"}).}

\item{language}{A character string specifying the language for the request. Defaults to \code{"en"}.}

\item{progress}{A logical value indicating whether to show progress messages during the data download and parsing. Defaults to \code{TRUE}.}
}
\value{
A tibble containing the indicator data for the specified countries. The following columns are included:
\describe{
\item{indicator_id}{The ID of the indicator (e.g., "NY.GDP.PCAP.KD").}
\item{country_id}{The ISO 2-country code of the country for which the data was retrieved.}
\item{date}{The date of the indicator data (either a year or month depending on the request).}
\item{value}{The value of the indicator for the given country and date.}
}
}
\description{
This function retrieves indicator data from the World Bank API for a specified set of countries.
The user can specify an indicator, date range, and other options to tailor the request. The data
is processed and returned in a tidy format, including country, indicator, date, and value fields.
}
\details{
This function constructs a request URL for the World Bank API, retrieves the relevant data,
and processes the response into a tidy format. The user can optionally specify a date range, and the
function will handle requests for multiple pages if necessary. If the \code{progress} parameter is \code{TRUE},
messages will be displayed during the request and parsing process.
}
\examples{
download_indicator(c("MX", "CA", "US"), "NY.GDP.PCAP.KD")
download_indicator(c("MX", "CA", "US"), "DPANUSSPB", start_date = 2012, end_date = 2013)
download_indicator(c("MX", "CA", "US"), "DPANUSSPB", start_date = "2012M01", end_date = "2012M05")
download_indicator("all", "NY.GDP.PCAP.KD", progress = FALSE)

}
