% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supported_indicators.R
\name{list_supported_indicators}
\alias{list_supported_indicators}
\title{Download and process indicators data from the World Bank API}
\usage{
list_supported_indicators(progress = TRUE)
}
\arguments{
\item{progress}{A logical value indicating whether to show progress messages during the
parsing of indicators. Defaults to \code{TRUE}. If set to \code{FALSE}, no progress message is shown.}
}
\value{
A tibble containing processed indicator information. The following columns are included:
\describe{
\item{indicator_id}{The ID of the indicator.}
\item{indicator_name}{The name of the indicator.}
\item{unit}{The unit of measurement for the indicator.}
\item{source_id}{The ID of the source that provided the indicator.}
\item{source_value}{The name of the source that provided the indicator.}
\item{source_note}{Any additional notes from the source.}
\item{source_organization}{The organization providing the source.}
\item{topics}{A nested tibble containing the topics associated with the indicator, which includes topic_id and topic_value}
}
}
\description{
This function retrieves a comprehensive list of indicators from the World Bank API.
The data is processed to extract and organize relevant fields such as the indicator's
ID, name, unit, source information, and associated topics. The function supports
optional progress messages during the parsing of indicators.
}
\details{
The data is fetched in JSON format from the World Bank API and is processed
to handle missing or empty fields. Topics related to each indicator are nested in
a separate tibble within the main dataframe. If there are no topics for an indicator,
\code{NA} values are used.
}
\examples{
list_supported_indicators()

}
