% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supported_regions.R
\name{list_supported_regions}
\alias{list_supported_regions}
\title{List supported regions for the World Bank API}
\source{
https://api.worldbank.org/v2/region
}
\usage{
list_supported_regions(language = "en")
}
\arguments{
\item{language}{A character string specifying the language code for the API response (default is "en" for English).}
}
\value{
A tibble with several columns, including:
\describe{
\item{id}{The unique identifier for the region (non-missing for 7 entries).}
\item{other_columns}{Other columns returned by the API, depending on the structure of the data.}
}
}
\description{
This function returns a tibble of supported regions for querying the World Bank API.
The regions include various geographic areas covered by the World Bank's datasets.
}
\details{
This function provides a reference for the supported regions, which are important for refining
queries related to geographic data in the World Bank's datasets. The \code{id} column is unique for seven key regions.
}
\examples{
# List all supported regions
list_supported_regions()

}
