% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_request.R
\name{perform_request}
\alias{perform_request}
\title{Internal function to perform API requests to the World Bank API}
\usage{
perform_request(
  resource,
  language = NULL,
  base_url = "https://api.worldbank.org/v2/"
)
}
\arguments{
\item{resource}{A character string specifying the resource endpoint for the World Bank API (e.g., "incomeLevels", "lendingTypes").}

\item{language}{A character string specifying the language code for the API response (default is "en" for English).}

\item{base_url}{A character string specifying the base URL of the World Bank API (default is "https://api.worldbank.org/v2").}
}
\value{
The raw response object from the API request.
}
\description{
This function constructs and sends a request to the specified World Bank API resource,
with the option to specify the language of the response.
}
\details{
This is an internal helper function used by other functions to interact with the World Bank API.
It constructs the request URL using the base URL, language, and resource, and then performs the request.
}
\keyword{internal}
