% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbwdi-package.R
\docType{package}
\name{wbwdi-package}
\alias{wbwdi}
\alias{wbwdi-package}
\title{wbwdi: Download World Development Indicators from the World Bank Development Indicators API}
\description{
A modern, flexible interface for accessing the World Bank’s World Development Indicators (WDI) API \url{https://datahelpdesk.worldbank.org/knowledgebase/articles/889392-about-the-indicators-api-documentation}. 'wbwdi' allows users to download, process, and analyze indicator data for multiple countries and years. 'wbwdi' relies on 'httr2' for multi-page request and error handling and using progress bars to keep users informed about the data processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidy-intelligence/r-wbwdi}
  \item Report bugs at \url{https://github.com/tidy-intelligence/r-wbwdi/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph.scheuch@gmail.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID})

}
\keyword{internal}
