% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdi_get_geographies.R
\name{wdi_get_geographies}
\alias{wdi_get_geographies}
\title{Download all geographies and regions from the World Bank API}
\usage{
wdi_get_geographies(language = "en", per_page = 1000)
}
\arguments{
\item{language}{A character string specifying the language for the API response. Defaults to \code{"en"} (English).
Other supported options include \code{"es"} (Spanish), \code{"fr"} (French), and others depending on the API.}

\item{per_page}{An integer specifying the number of records to fetch per request. Defaults to \code{1000}.}
}
\value{
A tibble with the following columns:
\describe{
\item{geography_id}{Character string representing the geography's unique identifier.}
\item{geography_iso2code}{Character string for the ISO2 country code.}
\item{geography_name}{Character string for the name of the geography.}
\item{region_id}{Character string representing the region's unique identifier.}
\item{region_iso2code}{Character string for the ISO2 region code.}
\item{region_name}{Character string for the name of the region.}
\item{admin_region_id}{Character string representing the administrative region's unique identifier.}
\item{admin_region_iso2code}{Character string for the ISO2 code of the administrative region.}
\item{admin_region_name}{Character string for the name of the administrative region.}
\item{lending_type_id}{Character string representing the lending type's unique identifier.}
\item{lending_type_iso2code}{Character string for the ISO2 code of the lending type.}
\item{lending_type_name}{Character string for the name of the lending type.}
\item{capital_city}{Character string for the name of the capital city.}
\item{longitude}{Numeric value for the longitude of the geography.}
\item{latitude}{Numeric value for the latitude of the geography.}
}
}
\description{
This function retrieves information about geographies (countries or regions) from the World Bank API.
It returns a tibble containing various details such as the geography's ID, ISO2 code, name, region information,
lending type, capital city, and coordinates.
}
\details{
This function sends a request to the World Bank API to retrieve data for all supported geographies
in the specified language. The data is then processed into a tidy format and includes information about the
country, such as its ISO code, capital city, geographical coordinates, and additional nested metadata about
regions, income levels, and lending types.
}
\examples{
# Download all geographies in English
wdi_get_geographies(language = "en")

# Download all geographies in Spanish
wdi_get_geographies(language = "zh")

}
