% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbwdi-package.R
\docType{package}
\name{wbwdi-package}
\alias{wbwdi}
\alias{wbwdi-package}
\title{wbwdi: Seamless Access to World Bank World Development Indicators (WDI)}
\description{
Access and analyze the World Bank’s World Development Indicators (WDI) using the corresponding API \url{https://datahelpdesk.worldbank.org/knowledgebase/articles/889392-about-the-indicators-api-documentation}. WDI provides more than 24,000 country or region-level indicators for various contexts. 'wbwdi' enables users to download, process and work with WDI series across multiple countries, aggregates, and time periods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidy-intelligence/r-wbwdi}
  \item \url{https://tidy-intelligence.github.io/r-wbwdi/}
  \item Report bugs at \url{https://github.com/tidy-intelligence/r-wbwdi/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph@tidy-intelligence.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID}) [copyright holder]

}
\keyword{internal}
